fw24_600_Day5

+++ backend/Routes/Code.routes.js
@@ -8,26 +8,29 @@ const codeRouter = Router();
 codeRouter.post("/upload", async (req, res) => {
     const userID=req.body.userID ;
     const userName=req.body.user;
-    const {title,content,language} =req.body;
+    const {title,content,language,about} =req.body;
     const payload={
         userID,
-        username:userName,
-        title,
+        owner:userName,
+        title:`${userName}/${title}`,
         content,
-        language
+        language,
+        about
     }
     try {
         const code = new CodeModel(payload);
         await code.save();
-        res.status(200).json({ message: 'Code Uploaded Sucessfull' })
+        res.status(200).json({ message: 'Code Uploaded Sucessfull',code })
     } catch (error) {
         res.status(400).json({ error: error.message })
     }
 });
 
 codeRouter.get("/getcode",async (req, res) => {
+    const userID=req.body.userID ;
+    const userName=req.body.user;
     try {
-        const data = await CodeModel.find();
+        const data = await CodeModel.find({userID});
         res.status(200).json({ message: 'Data Get Sucessfull',data })
     } catch (error) {
         res.status(400).json({ error: error.message })
@@ -35,9 +38,12 @@ codeRouter.get("/getcode",async (req, res) => {
 });
 codeRouter.patch("/update/:CodeId",async (req, res) => {
 
+    const userID=req.body.userID ;
+    const userName=req.body.user;
+
     const {CodeId} = req.params;
     try {
-        const data = await CodeModel.findByIdAndUpdate({_id:CodeId},req.body);
+        const data = await CodeModel.findByIdAndUpdate({_id:CodeId,userID},req.body);
         res.status(200).json({ message: 'Code Updated Sucessfull',data })
     } catch (error) {
         res.status(400).json({ error: error.message })
@@ -45,9 +51,11 @@ codeRouter.patch("/update/:CodeId",async (req, res) => {
 });
 
 codeRouter.delete("/delete/:CodeId",async (req, res) => {
+    const userID=req.body.userID ;
+    const userName=req.body.user;
     const {CodeId} = req.params;
     try {
-        const data = await CodeModel.findByIdAndDelete({_id:CodeId});
+        const data = await CodeModel.findByIdAndDelete({_id:CodeId,userID});
         res.status(200).json({ message: 'Code Deleted Sucessfull',data })
     } catch (error) {
         res.status(400).json({ error: error.message })
